requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.1
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0.1
django-debug-toolbar==6.1.0
django-debug-toolbar-force==0.2
django-filter==25.2
django-nine==0.2.7
djangorestframework==3.16.1
idna==3.11
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
urllib3==2.6.3
